+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+* add `--mqtt-keys name-only` to publish MQTT JSON keys without appending `|unit` (default remains `name-with-unit`)
+
 ## [2.0.0](https://github.com/yoch/modpoll2mqtt/compare/v1.6.0...v2.0.0) (2026-06-10)
 
 ### Project